refactor singleRowDeckMinin --> tableSingleRowMinin.module.scss
