+++ src/pagesMinin/TablesMinin/DeckCardsTableMinin/SingleRowCardMinin.tsx
@@ -6,7 +6,7 @@ import Typography from '@/components/ui/Typography/Typography'
 import { Button } from '@/components/ui/button'
 import { Table } from '@/components/ui/table'
 
-import s from '@/pagesMinin/TablesMinin/singleRowDeckMinin.module.scss'
+import s from '@/pagesMinin/TablesMinin/tableSingleRowMinin.module.scss'
 
 import { Card } from '../../../../services/decks/deck.types'
