+++ src/pagesMinin/TablesMinin/tableSingleRowMinin.module.scss
